+++ db/migrate/20181122141338_internationalise_post_title.rb
@@ -1,15 +1,10 @@
-class InternationalisePostTitle < ActiveRecord::Migration[5.2]
+class InternationalisePostTitle < ActiveRecord::Migration[5.0]
   def change
-    add_column :posts, :title_en, :string
-    add_column :posts, :title_nl, :string
-
-    Post.all.find_each do |p|
-      p.update(title_en: p.title, title_nl: p.title)
-    end
-
-    change_column :posts, :title_en, :string, null: false
-    change_column :posts, :title_nl, :string, null: false
-    remove_index :posts, [:title, :release]
-    remove_column :posts, :title
+    add_column :posts, :title_en, :string unless column_exists?(:posts, :title_en)
+    add_column :posts, :title_nl, :string unless column_exists?(:posts, :title_nl)
+    remove_index :posts, [:title, :release] if index_exists?(:posts, [:title, :release])
+    remove_column :posts, :title if column_exists?(:posts, :title)
+    change_column :posts, :title_en, :string, null: false if column_exists?(:posts, :title_en)
+    change_column :posts, :title_nl, :string, null: false if column_exists?(:posts, :title_nl)
   end
 end
