+++ db/migrate/20190328092228_drop_submission_details_table.rb
@@ -1,5 +1,5 @@
-class DropSubmissionDetailsTable < ActiveRecord::Migration[5.2]
+class DropSubmissionDetailsTable < ActiveRecord::Migration[5.0]
   def change
-    drop_table :submission_details
+    drop_table :submission_details if table_exists?(:submission_details)
   end
 end
