+++ db/migrate/20190328092326_remove_ugent_id_from_database.rb
@@ -1,5 +1,5 @@
-class RemoveUgentIdFromDatabase < ActiveRecord::Migration[5.2]
+class RemoveUgentIdFromDatabase < ActiveRecord::Migration[5.0]
   def change
-    remove_column :users, :ugent_id
+    remove_column :users, :ugent_id if column_exists?(:users, :ugent_id)
   end
 end
