+++ db/migrate/20190830135724_remove_exercise_token.rb
@@ -1,5 +1,5 @@
-class RemoveExerciseToken < ActiveRecord::Migration[6.0]
+class RemoveExerciseToken < ActiveRecord::Migration[5.0]
   def change
-    remove_column :exercises, :token
+    remove_column :exercises, :token if column_exists?(:exercises, :token)
   end
 end
