+++ db/migrate/20191003073541_remove_runner_column_from_judge.rb
@@ -1,5 +1,5 @@
-class RemoveRunnerColumnFromJudge < ActiveRecord::Migration[6.0]
+class RemoveRunnerColumnFromJudge < ActiveRecord::Migration[5.0]
   def change
-    remove_column :judges, :runner
+    remove_column :judges, :runner if column_exists?(:judges, :runner)
   end
 end
