+++ db/migrate/20191106142529_remove_posts.rb
@@ -1,9 +1,8 @@
-class RemovePosts < ActiveRecord::Migration[6.0]
+class RemovePosts < ActiveRecord::Migration[5.2]
   def change
-    drop_table :posts
-    drop_table :action_text_rich_texts
-    ActiveStorage::Attachment.find_each(&:purge)
-    drop_table :active_storage_attachments
-    drop_table :active_storage_blobs
+    drop_table :posts if table_exists?(:posts)
+    drop_table :action_text_rich_texts if table_exists?(:action_text_rich_texts)
+    drop_table :active_storage_attachments if table_exists?(:active_storage_attachments)
+    drop_table :active_storage_blobs if table_exists?(:active_storage_blobs)
   end
 end
